+++ components/feature/intent/src/test/java/mozilla/components/feature/intent/processing/TabIntentProcessorTest.kt
@@ -19,6 +19,7 @@ import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSession.LoadUrlFlags
+import mozilla.components.concept.storage.BookmarksStorage
 import mozilla.components.feature.search.SearchUseCases
 import mozilla.components.feature.session.SessionUseCases
 import mozilla.components.support.test.any
@@ -48,7 +49,10 @@ class TabIntentProcessorTest {
     private val session = mock<Session>()
     private val sessionUseCases = SessionUseCases(store, sessionManager)
     private val searchEngineManager = mock<SearchEngineManager>()
-    private val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager)
+    private val bookmarksStorage = mock<BookmarksStorage>()
+    private val searchUseCases = SearchUseCases(
+        testContext, store, searchEngineManager, sessionManager, bookmarksStorage
+    )
 
     @Before
     fun setup() {
@@ -255,7 +259,7 @@ class TabIntentProcessorTest {
         val engine = mock<Engine>()
         val sessionManager = spy(SessionManager(engine))
 
-        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager)
+        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager, bookmarksStorage)
         val sessionUseCases = SessionUseCases(store, sessionManager)
 
         val searchTerms = "mozilla android"
@@ -326,7 +330,7 @@ class TabIntentProcessorTest {
         val engine = mock<Engine>()
         val sessionManager = spy(SessionManager(engine))
 
-        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager)
+        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager, bookmarksStorage)
         val sessionUseCases = SessionUseCases(store, sessionManager)
 
         val searchTerms = "mozilla android"
@@ -385,7 +389,7 @@ class TabIntentProcessorTest {
         val engine = mock<Engine>()
         val sessionManager = spy(SessionManager(engine))
 
-        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager)
+        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager, bookmarksStorage)
         val sessionUseCases = SessionUseCases(store, sessionManager)
 
         val searchTerms = "mozilla android"