+++ components/feature/search/src/test/java/mozilla/components/feature/search/SearchUseCasesTest.kt
@@ -5,6 +5,7 @@
 package mozilla.components.feature.search
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import kotlinx.coroutines.runBlocking
 import mozilla.components.browser.search.SearchEngine
 import mozilla.components.browser.search.SearchEngineManager
 import mozilla.components.browser.session.Session
@@ -12,17 +13,21 @@ import mozilla.components.browser.session.SessionManager
 import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.state.SessionState
 import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.storage.BookmarksStorage
 import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.eq
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import mozilla.components.support.test.whenever
+import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.Mockito.timeout
+import org.mockito.Mockito.validateMockitoUsage
 import org.mockito.Mockito.verify
 
 @RunWith(AndroidJUnit4::class)
@@ -32,6 +37,7 @@ class SearchUseCasesTest {
     private lateinit var searchEngineManager: SearchEngineManager
     private lateinit var sessionManager: SessionManager
     private lateinit var store: BrowserStore
+    private lateinit var bookmarksStorage: BookmarksStorage
     private lateinit var useCases: SearchUseCases
 
     @Before
@@ -40,7 +46,8 @@ class SearchUseCasesTest {
         searchEngineManager = mock()
         sessionManager = mock()
         store = mock()
-        useCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager)
+        bookmarksStorage = mock()
+        useCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager, bookmarksStorage)
     }
 
     @Test
@@ -53,9 +60,9 @@ class SearchUseCasesTest {
         whenever(searchEngineManager.getDefaultSearchEngine(testContext)).thenReturn(searchEngine)
 
         useCases.defaultSearch(searchTerms, session)
-        assertEquals(searchTerms, session.searchTerms)
         val actionCaptor = argumentCaptor<EngineAction.LoadUrlAction>()
-        verify(store).dispatch(actionCaptor.capture())
+        verify(store, timeout(1000)).dispatch(actionCaptor.capture())
+        assertEquals(searchTerms, session.searchTerms)
         assertEquals(searchUrl, actionCaptor.value.url)
     }
 
@@ -69,7 +76,7 @@ class SearchUseCasesTest {
 
         useCases.newTabSearch(searchTerms, SessionState.Source.NEW_TAB)
         val actionCaptor = argumentCaptor<EngineAction.LoadUrlAction>()
-        verify(store).dispatch(actionCaptor.capture())
+        verify(store, timeout(1000)).dispatch(actionCaptor.capture())
         assertEquals(searchUrl, actionCaptor.value.url)
     }
 
@@ -82,15 +89,36 @@ class SearchUseCasesTest {
 
         var sessionCreatedForUrl: String? = null
 
-        val searchUseCases = SearchUseCases(testContext, store, searchEngineManager, sessionManager) { url ->
+        val searchUseCases = SearchUseCases(
+            testContext, store, searchEngineManager, sessionManager, bookmarksStorage
+        ) { url ->
             sessionCreatedForUrl = url
             Session(url).also { createdSession = it }
         }
 
         searchUseCases.defaultSearch("test")
+
+        val loadUrlActionCaptor = argumentCaptor<EngineAction.LoadUrlAction>()
+        verify(store, timeout(1000)).dispatch(loadUrlActionCaptor.capture())
+        assertEquals(createdSession!!.id, loadUrlActionCaptor.value.sessionId)
+        assertEquals(sessionCreatedForUrl, loadUrlActionCaptor.value.url)
         assertEquals("https://search.example.com", sessionCreatedForUrl)
         assertNotNull(createdSession!!)
-        verify(store).dispatch(EngineAction.LoadUrlAction(createdSession!!.id, sessionCreatedForUrl!!))
+    }
+
+    @Test
+    fun `DefaultSearchUseCase can load the url for a bookmark keyword`() = runBlocking {
+        val session = mock<Session>()
+        whenever(session.id).thenReturn("sessionId")
+        val bookmarkUrl = "https://search.example.com"
+        val bookmarkKeyword = "keyword"
+        whenever(bookmarksStorage.getBookmarkUrlForKeyword("keyword")).thenReturn(bookmarkUrl)
+
+        useCases.defaultSearch(bookmarkKeyword, session)
+
+        val loadUrlActionCaptor = argumentCaptor<EngineAction.LoadUrlAction>()
+        verify(store, timeout(1000)).dispatch(loadUrlActionCaptor.capture())
+        assertEquals(bookmarkUrl, loadUrlActionCaptor.value.url)
     }
 
     @Test
@@ -103,10 +131,10 @@ class SearchUseCasesTest {
         useCases.newPrivateTabSearch.invoke(searchTerms)
 
         val captor = argumentCaptor<Session>()
-        verify(sessionManager).add(captor.capture(), eq(true), eq(null), eq(null), eq(null))
+        verify(sessionManager, timeout(1000)).add(captor.capture(), eq(true), eq(null), eq(null), eq(null))
         assertTrue(captor.value.private)
         val actionCaptor = argumentCaptor<EngineAction.LoadUrlAction>()
-        verify(store).dispatch(actionCaptor.capture())
+        verify(store, timeout(1000)).dispatch(actionCaptor.capture())
         assertEquals(searchUrl, actionCaptor.value.url)
     }
 
@@ -121,12 +149,12 @@ class SearchUseCasesTest {
         val parentSession = mock<Session>()
         useCases.newPrivateTabSearch.invoke(searchTerms, parentSession = parentSession)
 
-        val captor = argumentCaptor<Session>()
-        verify(sessionManager).add(captor.capture(), eq(true), eq(null), eq(null), eq(parentSession))
-        assertTrue(captor.value.private)
-
         val actionCaptor = argumentCaptor<EngineAction.LoadUrlAction>()
-        verify(store).dispatch(actionCaptor.capture())
+        verify(store, timeout(1000)).dispatch(actionCaptor.capture())
         assertEquals(searchUrl, actionCaptor.value.url)
+
+        val captor = argumentCaptor<Session>()
+        verify(sessionManager, timeout(1000)).add(captor.capture(), eq(true), eq(null), eq(null), eq(parentSession))
+        assertTrue(captor.value.private)
     }
 }